refactor(chat): routes split in two files
